+++ test/core/test_operator_worker.py
@@ -6,20 +6,21 @@ def test_operator_info_request(operator):
     assert isinstance(operator.info(), _Info)
 
 
-def test_operator_compute_request(
+def test_operator_compute_german_request(
     operator,
     default_compute_input,
-    default_aoi,
+    default_german_aoi,
     default_aoi_properties,
     compute_resources,
+    mock_query_census_tables,
 ):
     computed_artifacts = operator.compute(
         resources=compute_resources,
-        aoi=default_aoi,
+        aoi=default_german_aoi,
         aoi_properties=default_aoi_properties,
         params=default_compute_input,
     )
 
-    assert len(computed_artifacts) == 1
+    assert len(computed_artifacts) == 2
     for artifact in computed_artifacts:
         assert isinstance(artifact, _Artifact)